Simplification du composant Home
