+++ src/app/index.tsx
@@ -4,13 +4,11 @@ import "./global.css";
 import { BottomSheetModalProvider } from "@gorhom/bottom-sheet";
 import * as Sentry from "@sentry/react-native";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import * as Clipboard from "expo-clipboard";
 import * as Notifications from "expo-notifications";
 import { Provider } from "jotai";
 import { useEffect, useState } from "react";
 import { View } from "react-native";
 import { GestureHandlerRootView } from "react-native-gesture-handler";
-// import { DevToolsBubble } from "react-native-react-query-devtools";
 import { Text } from "@/components/common/Text";
 import { ToastProvider } from "@/components/common/Toast";
 import { apiEnv } from "@/config";
@@ -31,15 +29,6 @@ Notifications.setNotificationHandler({
 const App = () => {
   const queryClient = new QueryClient();
 
-  const _onCopy = async (text: string) => {
-    try {
-      await Clipboard.setStringAsync(text);
-      return true;
-    } catch {
-      return false;
-    }
-  };
-
   const [isDevServerSelected, setIsDevServerSelected] = useState(false);
   useEffect(() => {
     const fetchIsDevServerSelected = async () => {
@@ -50,8 +39,6 @@ const App = () => {
     };
     fetchIsDevServerSelected();
   }, []);
-
-  // const isDev = process.env.NODE_ENV === "development";
 
   return (
     <Provider>
@@ -71,9 +58,6 @@ const App = () => {
                 </ToastProvider>
               </AuthProvider>
             </BottomSheetModalProvider>
-            {/* {isDev && (
-              <DevToolsBubble onCopy={_onCopy} queryClient={queryClient} />
-            )} */}
           </QueryClientProvider>
         </GestureHandlerRootView>
       </ThemeProvider>
