+++ src/app/navigation/RootNavigator.tsx
@@ -4,6 +4,7 @@ import { AppNavigator } from "@/app/navigation/AppNavigator";
 import { AuthNavigator } from "@/app/navigation/AuthNavigator";
 import { HomeLoading } from "@/app/screens/home/Home";
 import { useAuth } from "@/hooks/account/useAuth";
+import { usePushNotifications } from "@/hooks/home";
 import { i18nInitializedPromise } from "@/i18n";
 import { screenOptions } from "@/navigation/navigationConfig";
 import type { RootStackParamList } from "@/types";
@@ -12,6 +13,8 @@ const Stack = createStackNavigator<RootStackParamList>();
 
 export const RootNavigator = () => {
   const { user } = useAuth();
+  usePushNotifications();
+
   const [isI18nReady, setIsI18nReady] = useState(false);
 
   useEffect(() => {
