--- src/hooks/useRefreshOnFocus.ts
@@ -1,17 +1,0 @@
-import { useFocusEffect } from "@react-navigation/native";
-import { useCallback, useRef } from "react";
-
-export const useRefreshOnFocus = <T>(refetch: () => Promise<T>) => {
-  const firstTimeRef = useRef(true);
-
-  useFocusEffect(
-    useCallback(() => {
-      if (firstTimeRef.current) {
-        firstTimeRef.current = false;
-        return;
-      }
-
-      refetch();
-    }, [refetch]),
-  );
-};
